+++ src/pyramid/connector/discord/commands/api/register.py
@@ -1,11 +1,10 @@
 
 import logging
 from discord.ext.commands import Bot
-from pyramid.connector.discord.commands.api.abstract_command import AbstractCommand
-from pyramid.connector.discord.commands.api.parameters_command import ParametersCommand
+from pyramid.connector.discord.commands.api.abc import AbstractCommand
+from pyramid.connector.discord.commands.api.parameters import ParametersCommand
 
 COMMANDS_AUTOREGISTRED: dict[type[AbstractCommand], ParametersCommand] = {}
-
 
 def register_commands(bot: Bot, logger: logging.Logger, command_prefix: str | None = None):
 	for cls, parameters in COMMANDS_AUTOREGISTRED.items():
