+++ src/pyramid/connector/discord/commands/ping_command.py
@@ -1,8 +1,8 @@
 import math
 from discord import Interaction
-from pyramid.connector.discord.commands.api.abstract_command import AbstractCommand
-from pyramid.connector.discord.commands.api.annotation_command import discord_command
-from pyramid.connector.discord.commands.api.parameters_command import ParametersCommand
+from pyramid.connector.discord.commands.api.abc import AbstractCommand
+from pyramid.connector.discord.commands.api.annotation import discord_command
+from pyramid.connector.discord.commands.api.parameters import ParametersCommand
 
 @discord_command(parameters=ParametersCommand(description="Displays response time between bot and Discord API"))
 class PingCommand(AbstractCommand):
